feat: implement intern and management dashboard views with task tracking and reporting features and made components modular
