+++ src/app/(dashboard)/dashboard/intern/tasks/page.tsx
@@ -12,7 +12,6 @@ import {
 } from "lucide-react";
 import { useState } from "react";
 import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
 import {
   Card,
   CardContent,
