+++ src/app/(dashboard)/dashboard/intern/timesheet/page.tsx
@@ -5,7 +5,6 @@ import {
   Clock,
   FileText,
   Flame,
-  Gem,
   Loader2,
   Sparkles,
   Zap,
